Rust 1.66 clippy

+++ dolby_vision/src/c_structs/buffers.rs
@@ -232,31 +232,27 @@ impl<const N: usize> From<Option<[u64; N]>> for U64Data {
 
 impl Freeable for Data {
     unsafe fn free(&self) {
-        Vec::from_raw_parts(self.data as *mut u8, self.len as usize, self.len as usize);
+        Vec::from_raw_parts(self.data as *mut u8, self.len, self.len);
     }
 }
 
 impl Freeable for U64Data {
     unsafe fn free(&self) {
         if !self.data.is_null() {
-            Vec::from_raw_parts(self.data as *mut u64, self.len as usize, self.len as usize);
+            Vec::from_raw_parts(self.data as *mut u64, self.len, self.len);
         }
     }
 }
 
 impl Freeable for I64Data {
     unsafe fn free(&self) {
-        Vec::from_raw_parts(self.data as *mut i64, self.len as usize, self.len as usize);
+        Vec::from_raw_parts(self.data as *mut i64, self.len, self.len);
     }
 }
 
 impl Freeable for Data2D {
     unsafe fn free(&self) {
-        let list = Vec::from_raw_parts(
-            self.list as *mut *const Data,
-            self.len as usize,
-            self.len as usize,
-        );
+        let list = Vec::from_raw_parts(self.list as *mut *const Data, self.len, self.len);
 
         for data_ptr in list {
             let data = Box::from_raw(data_ptr as *mut Data);
@@ -267,11 +263,7 @@ impl Freeable for Data2D {
 
 impl Freeable for U64Data2D {
     unsafe fn free(&self) {
-        let list = Vec::from_raw_parts(
-            self.list as *mut *const U64Data,
-            self.len as usize,
-            self.len as usize,
-        );
+        let list = Vec::from_raw_parts(self.list as *mut *const U64Data, self.len, self.len);
 
         for data_ptr in list {
             let data = Box::from_raw(data_ptr as *mut U64Data);
@@ -282,11 +274,7 @@ impl Freeable for U64Data2D {
 
 impl Freeable for I64Data2D {
     unsafe fn free(&self) {
-        let list = Vec::from_raw_parts(
-            self.list as *mut *const I64Data,
-            self.len as usize,
-            self.len as usize,
-        );
+        let list = Vec::from_raw_parts(self.list as *mut *const I64Data, self.len, self.len);
 
         for data_ptr in list {
             let data = Box::from_raw(data_ptr as *mut I64Data);
@@ -297,11 +285,7 @@ impl Freeable for I64Data2D {
 
 impl Freeable for U64Data3D {
     unsafe fn free(&self) {
-        let list = Vec::from_raw_parts(
-            self.list as *mut *const U64Data2D,
-            self.len as usize,
-            self.len as usize,
-        );
+        let list = Vec::from_raw_parts(self.list as *mut *const U64Data2D, self.len, self.len);
 
         for data2d_ptr in list {
             let data2d = Box::from_raw(data2d_ptr as *mut U64Data2D);
@@ -312,11 +296,7 @@ impl Freeable for U64Data3D {
 
 impl Freeable for I64Data3D {
     unsafe fn free(&self) {
-        let list = Vec::from_raw_parts(
-            self.list as *mut *const I64Data2D,
-            self.len as usize,
-            self.len as usize,
-        );
+        let list = Vec::from_raw_parts(self.list as *mut *const I64Data2D, self.len, self.len);
 
         for data2d_ptr in list {
             let data2d = Box::from_raw(data2d_ptr as *mut I64Data2D);
